backend/requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6 
sqlalchemy==2.0.32 
psycopg2-binary==2.9.9 
pydantic==2.8.2 
python-dotenv==1.0.1 
python-jwt==4.1.0 
web3==7.2.0 
httpx==0.27.2 
pytest==8.3.2 
pytest-asyncio==0.23.8
